Consolidate some of the Dynamo locking fixtures
